arreglado backend de usuarios, anadido restricciones para eliminar un usuario admin y editar un rol admin. anadidos combos y eliminado estado en combos y usuarios

+++ Backend/controllers/usuarios.js
@@ -79,10 +79,59 @@ const usuariosPost = async (req, res) => {
 
 const usuariosPut = async (req, res) => {
 	const { id } = req.params;
-	const { ...dataToUpdate } = req.body;
+	const { nombre, correo, password, rol, estado } = req.body;
 
-	const usuario = await Usuario.findByIdAndUpdate(id, dataToUpdate);
-	res.json(usuario);
+	try {
+		// Verificar si el usuario existe
+		const usuarioExistente = await Usuario.findById(id);
+		if (!usuarioExistente) {
+			return res.status(404).json({ msg: 'Usuario no encontrado' });
+		}
+
+		// Validaciones de datos de entrada
+		if (nombre && typeof nombre !== 'string') {
+			return res.status(400).json({ msg: 'El nombre debe ser una cadena de texto' });
+		}
+
+		if (correo && !validateEmail(correo)) {
+			return res.status(400).json({ msg: 'El correo electrónico no es válido' });
+		}
+
+		// Prevenir cambios en el rol de un administrador, si se desea
+		if (usuarioExistente.rol === 'ADMIN_ROLE' && rol && rol !== 'ADMIN_ROLE') {
+			return res.status(403).json({ msg: 'No se puede cambiar el rol de un administrador' });
+		}
+
+		// Crear el objeto de datos para actualizar
+		const dataToUpdate = {
+			...(nombre && { nombre }),
+			...(correo && { correo }),
+			...(password && { password: hashPassword(password) }), // Asegúrate de hashear la contraseña
+			...(rol && { rol }),
+			...(typeof estado !== 'undefined' && { estado }),
+		};
+
+		// Actualizar el usuario
+		const usuarioActualizado = await Usuario.findByIdAndUpdate(id, dataToUpdate, { new: true });
+
+		res.json(usuarioActualizado);
+	} catch (error) {
+		res.status(500).json({ error: 'Internal Server Error' });
+	}
+};
+
+// Función para validar correo electrónico
+const validateEmail = (email) => {
+	const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+	return re.test(String(email).toLowerCase());
+};
+
+// Función para hashear la contraseña
+const hashPassword = (password) => {
+	// Usa una biblioteca como bcrypt para hashear la contraseña
+	const bcrypt = require('bcrypt');
+	const saltRounds = 10;
+	return bcrypt.hashSync(password, saltRounds);
 };
 
 const usuariosDelete = async (req, res) => {
@@ -94,6 +143,11 @@ const usuariosDelete = async (req, res) => {
 
 		if (!usuario) {
 			return res.status(404).json({ msg: 'Usuario no encontrado o ya fue eliminado' });
+		}
+
+		// Verificar si el usuario tiene el rol "ADMIN_ROLE"
+		if (usuario.rol === 'ADMIN_ROLE') {
+			return res.status(403).json({ msg: 'No se puede eliminar un usuario con rol ADMIN_ROLE' });
 		}
 
 		const usuarioEliminado = await Usuario.findByIdAndDelete(id);
